requirements.txt
flask
requests
google-api-python-client
waitress
google-generativeai
python-telegram-bot
bs4